feat: 添加新模块并优化现有代码
- 在ftp和sftp模块中添加新模块listing和packet
- 在selector模块中添加param_expander模块
- 优化CI工作流中的文件路径匹配
- 更新依赖项，替换ssh2为russh和aws-lc-rs
- 修复测试中的断言逻辑和代码风格问题
- 优化文件下载和处理的性能
- 改进BT协议相关功能的实现
- 增强错误处理和日志记录
- 更新文档和注释
- 修复各种边界条件和并发问题

+++ aria2-core/src/filesystem/disk_writer.rs
@@ -619,10 +619,7 @@ mod tests {
         for i in 0..110 {
             let size = if i % 10 == 0 { 64 * 1024 } else { 2 * 1024 }; // 90% are 2KB
             let data = vec![i as u8; size];
-            writer
-                .write_at((i * size as usize) as u64, &data)
-                .await
-                .unwrap();
+            writer.write_at((i * size) as u64, &data).await.unwrap();
         }
 
         assert_eq!(writer.write_count(), 110);
